+++ Reach/src/utopia/reach/component/label/text/TextLabel.scala
@@ -2,17 +2,16 @@ package utopia.reach.component.label.text
 
 import utopia.firmament.component.text.TextComponent
 import utopia.firmament.context.TextContext
-import utopia.firmament.drawing.immutable.{BackgroundDrawer, TextDrawer}
+import utopia.firmament.drawing.immutable.{CustomDrawableFactory, TextDrawer}
 import utopia.firmament.drawing.template.CustomDrawer
 import utopia.firmament.localization.LocalizedString
 import utopia.firmament.model.TextDrawContext
 import utopia.firmament.model.stack.StackInsets
 import utopia.genesis.text.Font
-import utopia.paradigm.color.ColorLevel.Standard
-import utopia.paradigm.color.{Color, ColorLevel, ColorRole}
+import utopia.paradigm.color.{Color, ColorRole}
 import utopia.paradigm.enumeration.Alignment
 import utopia.reach.component.factory.ComponentFactoryFactory.Cff
-import utopia.reach.component.factory.{FromContextFactory, TextContextualFactory}
+import utopia.reach.component.factory.{ContextualBackgroundAssignableFactory, FromContextFactory, TextContextualFactory}
 import utopia.reach.component.hierarchy.ComponentHierarchy
 import utopia.reach.component.template.CustomDrawReachComponent
 
@@ -57,14 +56,27 @@ case class TextLabelFactory(parentHierarchy: ComponentHierarchy)
 			betweenLinesMargin, allowLineBreaks), customDrawers, allowTextShrink)
 }
 
-case class ContextualTextLabelFactory(factory: TextLabelFactory, context: TextContext)
+case class ContextualTextLabelFactory(factory: TextLabelFactory, context: TextContext,
+                                      customDrawers: Vector[CustomDrawer] = Vector(), isHint: Boolean = false)
 	extends TextContextualFactory[ContextualTextLabelFactory]
+		with ContextualBackgroundAssignableFactory[TextContext, ContextualTextLabelFactory]
+		with CustomDrawableFactory[ContextualTextLabelFactory]
 {
+	// COMPUTED ---------------------------------
+	
+	/**
+	  * @return Copy of this factory that creates hint labels
+	  */
+	def hint = copy(isHint = true)
+	
+	
 	// IMPLEMENTED	-----------------------------
 	
 	override def self: ContextualTextLabelFactory = this
 	
-	override def withContext(newContext: TextContext) = ContextualTextLabelFactory(factory, newContext)
+	override def withCustomDrawers(drawers: Vector[CustomDrawer]): ContextualTextLabelFactory =
+		copy(customDrawers = drawers)
+	override def withContext(newContext: TextContext) = copy(context = newContext)
 	
 	
 	// OTHER	---------------------------------
@@ -72,11 +84,9 @@ case class ContextualTextLabelFactory(factory: TextLabelFactory, context: TextCo
 	/**
 	  * Creates a new text label utilizing contextual information
 	  * @param text Text displayed on this label
-	  * @param customDrawers Additional custom drawing (default = empty)
-	  * @param isHint Whether this label should be considered a hint (affects text color)
 	  * @return A new label
 	  */
-	def apply(text: LocalizedString, customDrawers: Seq[CustomDrawer] = Vector(), isHint: Boolean = false) =
+	def apply(text: LocalizedString) =
 		factory(text, context.font, if (isHint) context.hintTextColor else context.textColor,
 			context.textAlignment, context.textInsets, context.betweenLinesMargin.optimal, customDrawers,
 			context.allowLineBreaks, context.allowTextShrink)
@@ -85,27 +95,19 @@ case class ContextualTextLabelFactory(factory: TextLabelFactory, context: TextCo
 	  * Creates a new text label with solid background utilizing contextual information
 	  * @param text              Text displayed on this label
 	  * @param background        Label background color
-	  * @param customDrawers Additional custom drawing (default = empty)
-	  * @param isHint            Whether this label should be considered a hint (affects text color)
 	  * @return A new label
 	  */
-	def withCustomBackground(text: LocalizedString, background: Color,
-	                         customDrawers: Seq[CustomDrawer] = Vector(), isHint: Boolean = false) =
-		mapContext { _.against(background) }(text,
-			BackgroundDrawer(background) +: customDrawers, isHint)
+	@deprecated("Replaced with .withBackground(Color).apply(LocalizedString)", "v1.1")
+	def withCustomBackground(text: LocalizedString, background: Color) = withBackground(background)(text)
 	
 	/**
 	  * Creates a new text label with solid background utilizing contextual information
 	  * @param text              Text displayed on this label
 	  * @param role              Label background color role
-	  * @param preferredShade    Preferred color shade (default = standard)
-	  * @param customDrawers Additional custom drawing (default = empty)
-	  * @param isHint            Whether this label should be considered a hint (affects text color)
 	  * @return A new label
 	  */
-	def withBackground(text: LocalizedString, role: ColorRole, preferredShade: ColorLevel = Standard,
-	                   customDrawers: Seq[CustomDrawer] = Vector(), isHint: Boolean = false) =
-		withCustomBackground(text, context.color.preferring(preferredShade)(role), customDrawers, isHint)
+	@deprecated("Replaced with .withBackground(ColorRole).apply(LocalizedString)", "v1.1")
+	def withBackground(text: LocalizedString, role: ColorRole): TextLabel = withBackground(role).apply(text)
 }
 
 /**